[TASK] Split bookmark list and items
