+++ src/minimal/minimalApp/bookmarksItem.vue
@@ -1,0 +1,155 @@
+<template>
+  <div :title="prediction && prediction.text" class="mdl-cell mdl-cell--2-col mdl-cell--4-col-tablet mdl-cell--6-col-phone mdl-shadow--2dp mdl-grid bookEntry" style="position: relative; height: 250px; padding: 0; width: 210px; height: 293px;">
+    <div class="data title lazyBack init" :data-src="imageHi(item)" style="background-image: url(); background-color: grey; background-size: cover; background-position: center center; background-repeat: no-repeat; width: 100%; position: relative; padding-top: 5px;">
+
+      <div v-if="prediction && prediction.text" class="mdl-shadow--2dp" style=" position: absolute; top: 0; right: 0; background-color: rgba(255, 255, 255, 0.9); padding: 0px 5px; margin: 5px 0; text-align: center;">
+        {{preTexter(prediction)}}
+      </div>
+
+      <a :href="item.url" style="position: absolute; top: 0; left: 0; right: 0; bottom: 0;"></a>
+      <span class="mdl-shadow--2dp" style="position: absolute; bottom: 0; display: block; background-color: rgba(255, 255, 255, 0.9); padding-top: 5px; display: inline-flex; align-items: center; justify-content: space-between; left: 0; right: 0; padding-right: 8px; padding-left: 8px; padding-bottom: 8px;">
+        <a :href="item.url" style="color: black; text-decoration: none;">
+          {{item.title}}
+        </a>
+        <div id="p1" class="mdl-progress" style="position: absolute; top: -4px; left: 0;">
+          <div class="progressbar bar bar1" :style="progress(item)"></div>
+          <div v-if="hasTotalEp(item)" class="bufferbar bar bar2" style="width: calc(100% + 1px);"></div>
+          <div v-if="prediction && prediction.tagEpisode" class="predictionbar bar kal-ep-pre" :style="predictionBar(item)"></div>
+          <div class="auxbar bar bar3" style="width: 0%;"></div>
+        </div>
+        <div class="data progress mdl-chip mdl-chip--contact mdl-color--indigo-100" style="float: right; line-height: 20px; height: 20px; padding-right: 4px; margin-left: 5px;">
+          <div class="link mdl-chip__contact mdl-color--primary mdl-color-text--white" style="line-height: 20px; height: 20px; margin-right: 0;">{{item.watchedEp}}</div>
+          <a class="mal-sync-stream" v-if="streamUrl(item)" :title="streamUrl(item).split('/')[2]" target="_blank" style="margin: 0 5px;" :href="streamUrl(item)">
+            <img :src="favicon(streamUrl(item).split('/')[2])">
+          </a>
+          <a v-if="continueUrl" class="nextStream" title="Continue watching" target="_blank" style="margin: 0 5px 0 0; color: #BABABA;" :href="continueUrl">
+            <img :src="assetUrl('double-arrow-16px.png')" width="16" height="16">
+          </a>
+          <a v-if="resumeUrl" class="resumeStream" title="Resume watching" target="_blank" style="margin: 0 5px 0 0; color: #BABABA;" :href="resumeUrl">
+            <img :src="assetUrl('arrow-16px.png')" width="16" height="16">
+          </a>
+        </div>
+      </span>
+    </div>
+  </div>
+</template>
+
+<script type="text/javascript">
+  import * as provider from "./../../provider/provider.ts";
+  export default {
+    data: function(){
+      return {
+        prediction: undefined,
+        resumeUrl: null,
+        continueUrl: null,
+      }
+    },
+    props: {
+      item: {
+        type: Object,
+      },
+    },
+    mounted: async function(){
+      if(typeof this.item.resume === 'undefined'){
+        var resumeUrl = null;
+        var continueUrl = null;
+        var id = utils.urlPart(this.item.url, 4);
+        var type = utils.urlPart(this.item.url, 3);
+        var resumeUrlObj = await utils.getResumeWaching(type, id);
+        var continueUrlObj = await utils.getContinueWaching(type, id);
+        var curEp = parseInt(this.item.watchedEp.toString());
+
+        con.log('Resume', resumeUrlObj, 'Continue', continueUrlObj);
+        if(typeof continueUrlObj !== 'undefined' && continueUrlObj.ep === (curEp+1)){
+          continueUrl = continueUrlObj.url;
+        }else if(typeof resumeUrlObj !== 'undefined' && resumeUrlObj.ep === curEp){
+          resumeUrl = resumeUrlObj.url;
+        }
+        this.resumeUrl = resumeUrl;
+        this.continueUrl = continueUrl;
+      }
+
+      if(typeof this.prediction === 'undefined'){
+        utils.epPredictionUI(utils.urlPart(this.item.url, 4), utils.urlPart(this.item.url, 3), (prediction) => {
+          this.prediction = prediction;
+        });
+      }
+    },
+    updated: function(){
+
+    },
+    methods: {
+      imageHi: function(item){
+        var imageHi = item.image;
+        var regexDimensions = /\/r\/\d*x\d*/g;
+        if ( regexDimensions.test(imageHi) ) {
+          imageHi = imageHi.replace(/v.jpg$/g, '.jpg').replace(regexDimensions, '');
+        }
+        return imageHi;
+      },
+      barTotal: function(item){
+
+        if(this.prediction && this.prediction.tagEpisode && !this.hasTotalEp(item)){
+          if(this.prediction.tagEpisode > item.watchedEp){
+            return Math.ceil(this.prediction.tagEpisode * 1.2);
+          }else{
+            return Math.ceil(item.watchedEp * 1.2);
+          }
+        }
+
+        return item.totalEp;
+      },
+      hasTotalEp: function(item){
+        return parseInt(item.totalEp) != 0
+      },
+      progress: function(item){
+        var width = ( item.watchedEp / this.barTotal(item) ) * 100;
+        return 'width: '+width+'%;'
+      },
+      predictionBar: function(item){
+        var predictionProgress = ( this.prediction.tagEpisode / this.barTotal(item) ) * 100;
+        var color = 'orange';
+        if(this.prediction.color != ''){
+          color = this.prediction.color;
+        }
+        return 'width: '+predictionProgress+'%; background-color: '+color;
+      },
+      streamUrl: function(item){
+        return utils.getUrlFromTags(item.tags);
+      },
+      favicon: function(domain){
+        return utils.favicon(domain);
+      },
+      assetUrl: function(asset){
+        return api.storage.assetUrl(asset);
+      },
+      preTexter: function(prediction){
+        var pre = prediction.prediction;
+        var diffDays = pre.diffDays;
+        var diffHours = pre.diffHours;
+        var diffMinutes = pre.diffMinutes;
+        //Round hours
+        if(diffDays > 1 && diffHours > 12){
+          diffDays++;
+        }
+
+        var text = '';
+        if(diffDays > 1){
+          return text+diffDays+' Days';
+        }
+        if(diffDays == 1){
+          text += diffDays+' Day ';
+        }
+
+        if(diffHours > 1){
+          return text+diffHours+' Hours';
+        }
+        if(diffHours == 1){
+          text += diffHours+' Hour ';
+        }
+
+        return text+diffMinutes+' mins';
+      },
+    }
+  }
+</script>
